Convert TIMESTAMP sql type to Date object
When using DuckDB TIMESTAMP data types, Arrow serializes to the int64 ms since epoch, but Mosaic never reconverts back into a Javascript Date object before passing the value to Plot. Existing behavior is that we see integers in the axis instead of formatted Date/time. Added a bit of a hack to the Marks class to catch this datatype but please let me know where the more appropriate place to include it is. Thank you!!